feat: 루틴의 MyRoutinePage 완성(아이콘 띄우기제외)

See merge request s13-mobility-smarthome-sub1/S13P21D208!136

+++ frontend/app/src/main/java/com/example/eeum/data/model/response/floorplans/FloorPlansList.kt
@@ -3,8 +3,8 @@ package com.example.eeum.data.model.response.floorplans
 data class FloorPlansList(
     val createdAt: String,
     val floorplanId: Int,
-    val floorplansX: Float,
-    val floorplansY: Float,
+    val floorplansX: Double,
+    val floorplansY: Double,
     val homeId: Int,
     val imageUrl: String,
     val square: Double,
